size_t is unsigned, yet in a few places size_t variables are checked to see if they are < 0.  This moves or casts to int as needed, without changing too much.

+++ openslp/common/slp_dhcp.c
@@ -395,7 +395,7 @@ int DHCPGetOptionInfo(unsigned char * dhcpOptCodes, int dhcpOptCodeCnt,
    uint8_t rcvbuf[512];
    struct hostent * hep;
    uint8_t * p;
-   size_t rcvbufsz = 0;
+   int rcvbufsz = 0;
    char host[256];
 
    /* Get our IP and MAC addresses */
